Move solver to service package
