+++ core/src/test/kotlin/org/evomaster/core/solver/ColumnTypeVocabularyTest.kt
@@ -5,7 +5,7 @@ import org.junit.jupiter.api.Assertions.assertFalse
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
 import org.evomaster.core.database.sql.solver.SmtLibGenerator
-import org.evomaster.core.database.sql.solver.SMTLibZ3DbConstraintSolver
+import org.evomaster.core.database.sql.solver.service.SMTLibZ3DbConstraintSolver
 
 /**
  * Two vocabularies describe the same column type: [SmtLibGenerator.TYPE_MAP], which decides the SMT-LIB
